+++ Faker_visa_credit_card_number_test.v
@@ -1,0 +1,11 @@
+module test
+
+import faker { Faker, is_valid_credit_card_number }
+
+fn test_returns_visa_credit_card_number() {
+    mut fake := Faker{}
+
+    credit_card_number := fake.visa_credit_card_number()
+
+    assert is_valid_credit_card_number(credit_card_number)
+}