+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - IP V4 fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
 - Company name fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
+- Visa credit card number fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
 
 ### Fixed